Add 메뉴 리포지토리 생성
메뉴 리포지 토리 및 인증 관련 설정 완료

+++ src/main/java/com/stick/rest/config/security/CustomUserDetailService.java
@@ -15,6 +15,6 @@ public class CustomUserDetailService implements UserDetailsService {
     private final UserJpaRepo userJpaRepo;
 
     public UserDetails loadUserByUsername(String userPk) {
-        return userJpaRepo.findById(Integer.valueOf(userPk)).orElseThrow(CUserNotFoundException::new);
+        return userJpaRepo.findById(Long.valueOf(userPk)).orElseThrow(CUserNotFoundException::new);
     }
 }
